migrate to use-highlight-labels hook. enabled via use-plan-upgradeability hook
